# synapse-server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend API for the Synapse Study Assistant.",
        "title": "Synapse API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/study/sessions": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Start a new study session for a deck in a specific mode (review, cram, quiz)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Study"
                ],
                "summary": "Start a study session",
                "parameters": [
                    {
                        "description": "Session Parameters",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler.startSessionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/study/sessions/{id}/answer": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Submit an answer rating for a card in an active study session",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Study"
                ],
                "summary": "Submit an answer",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Answer Details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler.answerRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/study/sessions/{id}/cards/next": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Fetch the next batch of due cards for an active session",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Study"
                ],
                "summary": "Get next due cards for session",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_synapse_server_internal_domain.Card"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/study/sessions/{id}/end": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "End an active study session and save progress",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Study"
                ],
                "summary": "End a study session",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/study/sessions/{id}/results": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve the scorecard and wrong answer details for a completed quiz session",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Study"
                ],
                "summary": "Get quiz results",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_synapse_server_internal_domain.QuizResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_synapse_server_internal_domain.Card": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deckID": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "dueAt": {
                    "description": "nil → card not yet studied (new)",
                    "type": "string"
                },
                "easiness": {
                    "description": "SM-2 E-factor (default 2.5)",
                    "type": "number",
                    "format": "float64"
                },
                "id": {
                    "type": "string"
                },
                "interval": {
                    "description": "SRS scheduling fields (initialised on first study)",
                    "type": "integer"
                },
                "repetitions": {
                    "description": "total successful reviews",
                    "type": "integer"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_synapse_server_internal_domain.Tag"
                    }
                },
                "type": {
                    "$ref": "#/definitions/github_com_synapse_server_internal_domain.CardType"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "github_com_synapse_server_internal_domain.CardType": {
            "type": "string",
            "enum": [
                "flashcard",
                "cloze",
                "free_response",
                "multiple_choice"
            ],
            "x-enum-varnames": [
                "CardTypeFlashcard",
                "CardTypeCloze",
                "CardTypeFreeResponse",
                "CardTypeMultipleChoice"
            ]
        },
...
        "github_com_synapse_server_internal_domain.QuizResult": {
            "type": "object",
            "properties": {
                "totalCorrect": {
                    "type": "integer"
                },
                "totalWrong": {
                    "type": "integer"
                },
                "wrongAnswers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_synapse_server_internal_domain.WrongAnswerSummary"
                    }
                }
            }
        },
...
        "github_com_synapse_server_internal_domain.StudyMode": {
            "type": "string",
            "enum": [
                "learn",
                "review",
                "cram",
                "quiz"
            ],
            "x-enum-varnames": [
                "StudyModeLearn",
                "StudyModeReview",
                "StudyModeCram",
                "StudyModeQuiz"
            ]
        },
        "github_com_synapse_server_internal_domain.Tag": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userID": {
                    "type": "string"
                }
            }
        },
        "github_com_synapse_server_internal_domain.WrongAnswerSummary": {
            "type": "object",
            "properties": {
                "cardId": {
                    "type": "string"
                },
                "correctBack": {
                    "type": "string"
                },
                "front": {
                    "type": "string"
                }
            }
        },
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
        "internal_handler.answerRequest": {
            "type": "object",
            "required": [
                "cardId",
                "rating",
                "timeTaken"
            ],
            "properties": {
                "cardId": {
                    "type": "string"
                },
                "rating": {
                    "type": "integer",
                    "maximum": 4,
                    "minimum": 1
                },
                "timeTaken": {
                    "type": "integer",
                    "minimum": 0
                }
            }
        },
...
        "internal_handler.startSessionRequest": {
            "type": "object",
            "required": [
                "deckId",
                "mode"
            ],
            "properties": {
                "deckId": {
                    "type": "string"
                },
                "mode": {
                    "$ref": "#/definitions/github_com_synapse_server_internal_domain.StudyMode"
                }
            }
        },
...
    }
}